# zentra/docs/swagger.yaml
basePath: /api
definitions:
...
  handlers.CreateTaskRequest:
    description: Create task request model
    properties:
      completed_at:
        example: "2024-03-24T21:41:49Z"
        type: string
      employee_id:
        example: 1
        type: integer
      notes:
        example: Handle with care
        type: string
      order_item_id:
        example: 1
        type: integer
      sequence_number:
        example: 1
        type: integer
      started_at:
        example: "2024-03-24T21:41:49Z"
        type: string
      status:
        example: pending
        type: string
      task_type:
        example: washing
        type: string
    required:
    - order_item_id
    - sequence_number
    - status
    - task_type
    type: object
...
  handlers.ErrorResponse:
    description: Standard error response model
    properties:
      error:
        example: Invalid request parameters
        type: string
    type: object
...
  handlers.SuccessResponse:
    description: Standard success response model
    properties:
      message:
        example: Operation completed successfully
        type: string
    type: object
...
  handlers.TaskResponse:
    description: Task response model
    properties:
      completed_at:
        example: "2024-03-24T21:41:49Z"
        type: string
      created_at:
        example: "2024-03-24T21:41:49Z"
        type: string
      created_by:
        example: admin
        type: string
      employee_id:
        example: 1
        type: integer
      employee_name:
        example: John Doe
        type: string
      id:
        example: 1
        type: integer
      notes:
        example: Handle with care
        type: string
      order_item_id:
        example: 1
        type: integer
      sequence_number:
        example: 1
        type: integer
      started_at:
        example: "2024-03-24T21:41:49Z"
        type: string
      status:
        example: pending
        type: string
      task_type:
        example: washing
        type: string
      tenant_id:
        example: 1
        type: integer
      updated_at:
        example: "2024-03-24T21:41:49Z"
        type: string
      updated_by:
        example: admin
        type: string
    type: object
...
  handlers.UpdateTaskRequest:
    description: Update task request model
    properties:
      completed_at:
        example: "2024-03-24T21:41:49Z"
        type: string
      employee_id:
        example: 1
        type: integer
      notes:
        example: Handle with care
        type: string
      order_item_id:
        example: 1
        type: integer
      sequence_number:
        example: 1
        type: integer
      started_at:
        example: "2024-03-24T21:41:49Z"
        type: string
      status:
        example: pending
        type: string
      task_type:
        example: washing
        type: string
    required:
    - order_item_id
    - sequence_number
    - status
    - task_type
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Vomo backend API documentation
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Vomo API
  version: "1.0"
paths:
...
  /tasks:
    get:
      description: Get all tasks
      parameters:
      - description: Tenant ID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.TaskResponse'
            type: array
        "400":
          description: Invalid request parameters
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get all tasks
      tags:
      - Task
    post:
      consumes:
      - application/json
      description: Create a new task with the provided details
      parameters:
      - description: Tenant ID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: Task Data
        in: body
        name: task
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateTaskRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.TaskResponse'
        "400":
          description: Invalid request parameters
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new task
      tags:
      - Task
  /tasks/{id}:
    delete:
      description: Delete an existing task
      parameters:
      - description: Tenant ID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.SuccessResponse'
        "400":
          description: Invalid request parameters
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Task not found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a task
      tags:
      - Task
    get:
      description: Get task details by ID
      parameters:
      - description: Tenant ID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.TaskResponse'
        "400":
          description: Invalid request parameters
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Task not found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get a task by ID
      tags:
      - Task
    put:
      consumes:
      - application/json
      description: Update an existing task with new details
      parameters:
      - description: Tenant ID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      - description: Task Data
        in: body
        name: task
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateTaskRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.TaskResponse'
        "400":
          description: Invalid request parameters
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Task not found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a task
      tags:
      - Task
  /tasks/by-employee:
    get:
      description: Get all tasks assigned to an employee
      parameters:
      - description: Tenant ID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: Employee ID
        in: query
        name: employee_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.TaskResponse'
            type: array
        "400":
          description: Invalid request parameters
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get tasks by employee ID
      tags:
      - Task
  /tasks/by-order-item:
    get:
      description: Get all tasks for a specific order item
      parameters:
      - description: Tenant ID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: Order Item ID
        in: query
        name: order_item_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.TaskResponse'
            type: array
        "400":
          description: Invalid request parameters
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get tasks by order item ID
      tags:
      - Task
  /tasks/by-status:
    get:
      description: Get all tasks with a specific status
      parameters:
      - description: Tenant ID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: Task Status
        in: query
        name: status
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.TaskResponse'
            type: array
        "400":
          description: Invalid request parameters
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get tasks by status
      tags:
      - Task
  /tasks/by-type:
    get:
      description: Get all tasks of a specific type
      parameters:
      - description: Tenant ID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: Task Type
        in: query
        name: task_type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.TaskResponse'
            type: array
        "400":
          description: Invalid request parameters
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get tasks by task type
      tags:
      - Task
...
produces:
- application/json
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"